Fix some platform types nullability and remove locale restrictions

+++ src/main/kotlin/com/github/couchtracker/server/infoProviders/ids/Tmdb.kt
@@ -4,18 +4,18 @@ import com.github.couchtracker.server.model.externalIds.TmdbExternalId
 
 
 sealed interface TmdbId {
-    val value: Int
+    val value: Long
 
     fun toExternalId() = TmdbExternalId(value)
 }
 
 @JvmInline
-value class TmdbShowId(override val value: Int) : TmdbId {
+value class TmdbShowId(override val value: Long) : TmdbId {
     init {
         requireTmdbId(value)
     }
 }
 
-private fun requireTmdbId(id: Int) {
+private fun requireTmdbId(id: Long) {
     require(id > 0)
 }
